Click note behavior complete

+++ core/src/com/kurante/projectvoice_gdx/util/extensions/Number.kt
@@ -15,7 +15,7 @@ fun Int.mapRange(fromMin: Int, fromMax: Int, toMin: Float, toMax: Float) =
     (this - fromMin) * (toMax - toMin) / (fromMax - fromMin).toFloat() + toMin
 
 fun Int.mapRange(fromMin: Int, fromMax: Int, toMin: Double, toMax: Double) =
-    (this - fromMin) * (toMax - toMin) / (fromMax - fromMin).toFloat() + toMin
+    (this - fromMin) * (toMax - toMin) / (fromMax - fromMin) + toMin
 
 // Assume fromMin = 0 and fromMax = 1f
 fun Float.mapRange(toMin: Float, toMax: Float, clamped: Boolean = true): Float {
